Refactor DraggableSwitch into a module
I also moved the utility functions into a separate `_utils.js` module.

I'm still unsure if this will result in a better/more optimized bundle
size. But from a theoretical perspective, `DraggableSwitch` is only a
progressive enhancement, not a necessity, to the base functionality of
the `<theme-switch>` component. Therefore, making it optional by moving
it into a separate module makes sense.